wip to see if async pubsub listener works
